Quick Match
A very simple and not too intelligence quick match has been
implemented

Given a list of people, it will provide all possible matches without
duplicates. This is for the entire season, which means, it does not
allow for new addition or cancellations.

The current implementation is also not very efficient. As it needs to
check for pair membership, meaning it doing double the work that it
needs to do.